Data.Recall

Description

A typeclass for generic memoization. Inspired by http://research.microsoft.com/en-us/um/people/simonpj/papers/assoc-types/fun-with-type-funs/typefun.pdf.

Memo Typeclass

class Memo a

Represents an approach to memoizing functions from a given type.

Minimal complete definition

toTable, fromTable

memoize :: Memo a => (a -> b) -> a -> b

Memoize a function according to its instance of Memo.

Generic memoization functions

asArray :: (Bounded a, Ix a) => (a -> b) -> a -> b

Memoize a function with a bounded domain into a boxed array.

asEnumTree :: (Bounded a, Enum a) => (a -> b) -> a -> b

Memoize a function with a boudned domain into a tree according to its domain's enum instance.

asTree :: (Bounded a, Integral a) => (a -> b) -> a -> b

Memoize a function with a bounded domain into a tree according to its domain's integral instance.